services/consumer_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
kafka-python==2.3.0
